requirements.txt
Flask==2.0.2
python-dotenv
flask_mongoengine==1.0.0